Seperate ModelStub class
