+++ src/test/java/seedu/address/logic/commands/EditPersonDescriptorTest.java
@@ -14,7 +14,7 @@ import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_HUSBAND;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.EditCommand.EditPersonDescriptor;
-import seedu.address.testutil.EditPersonDescriptorBuilder;
+import seedu.address.testutil.builders.EditPersonDescriptorBuilder;
 
 public class EditPersonDescriptorTest {
