+++ src/test/java/seedu/address/model/ModelManagerTest.java
@@ -18,7 +18,7 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.model.person.predicates.NameContainsKeywordsPredicate;
-import seedu.address.testutil.AddressBookBuilder;
+import seedu.address.testutil.builders.AddressBookBuilder;
 
 public class ModelManagerTest {
