+++ src/test/java/seedu/address/model/module/ModuleTest.java
@@ -15,7 +15,7 @@ import static seedu.address.testutil.TypicalPersons.IDA;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.model.person.Person;
-import seedu.address.testutil.ModuleBuilder;
+import seedu.address.testutil.builders.ModuleBuilder;
 
 public class ModuleTest {
     @Test
