+++ src/test/java/seedu/address/model/module/UniqueModuleListTest.java
@@ -16,7 +16,7 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.model.module.exceptions.DuplicateModuleException;
 import seedu.address.model.module.exceptions.ModuleNotFoundException;
-import seedu.address.testutil.ModuleBuilder;
+import seedu.address.testutil.builders.ModuleBuilder;
 
 public class UniqueModuleListTest {
     private final UniqueModuleList uniqueModuleList = new UniqueModuleList();
