+++ src/test/java/seedu/address/model/person/predicates/OfficeContainsKeywordsPredicateTest.java
@@ -11,7 +11,7 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.testutil.PersonBuilder;
+import seedu.address.testutil.builders.PersonBuilder;
 
 public class OfficeContainsKeywordsPredicateTest {
     @Test
