+++ src/test/java/seedu/address/testutil/builders/EditPersonDescriptorBuilder.java
@@ -1,4 +1,4 @@
-package seedu.address.testutil;
+package seedu.address.testutil.builders;
 
 import java.util.Set;
 import java.util.stream.Collectors;
