+++ src/test/java/seedu/address/testutil/builders/ModuleCodeBuilder.java
@@ -1,4 +1,4 @@
-package seedu.address.testutil;
+package seedu.address.testutil.builders;
 
 import seedu.address.model.module.ModuleCode;
